requirements.txt
# Telegram 监控独立项目依赖

# Telegram 客户端
telethon>=1.34.0

# AI 分析
google-genai>=1.0.0           # 新版 Gemini SDK

# 数据库
sqlalchemy>=2.0.0

# 配置管理
python-dotenv>=1.0.0

# 网络请求
requests>=2.31.0

# 邮件
markdown>=3.4.0
